+++ Contents/Code/__init__.py
@@ -201,6 +201,13 @@ class PhoenixAdultAgent(Agent.Movies):
                     results = PAsearchSites.networkGammaEnt.search(results, encodedTitle, title, searchTitle, siteNum, lang, searchByDateActor, searchDate, searchSiteID)
 
             ###############
+            ## Blowpass/My XXX Pass
+            ###############
+            if siteNum == 861:
+                if searchSiteID == 9999 or searchSiteID == 861:
+                    results = PAsearchSites.networkGammaEnt.search(results, encodedTitle, title, searchTitle, siteNum, lang, searchByDateActor, searchDate, searchSiteID)
+
+            ###############
             ## Mile High Media
             ###############
             if siteNum == 361:
@@ -1187,6 +1194,48 @@ class PhoenixAdultAgent(Agent.Movies):
                 if searchSiteID == 9999 or searchSiteID == 843:
                     results = PAsearchSites.networkGammaEnt.search(results, encodedTitle, title, searchTitle, siteNum, lang, searchByDateActor, searchDate, searchSiteID)
 
+            ###############
+            ## Deviant Hardcore
+            ###############
+            if siteNum == 859:
+                if searchSiteID == 9999 or searchSiteID == 859:
+                    results = PAsearchSites.network1service.search(results, encodedTitle, title, searchTitle, siteNum, lang, searchByDateActor, searchDate, searchSiteID)
+
+            ###############
+            ## She Will Cheat
+            ###############
+            if siteNum == 860:
+                if searchSiteID == 9999 or searchSiteID == 860:
+                    results = PAsearchSites.network1service.search(results, encodedTitle, title, searchTitle, siteNum, lang, searchByDateActor, searchDate, searchSiteID)
+
+            ###############
+            ## SinsLife
+            ###############
+            if siteNum == 862:
+                if searchSiteID == 9999 or searchSiteID == 862:
+                    results = PAsearchSites.siteSinsLife.search(results, encodedTitle, title, searchTitle, siteNum, lang, searchByDateActor, searchDate, searchSiteID)
+
+            ###############
+            ## Puffy Network
+            ###############
+            if siteNum == 863:
+                if searchSiteID == 9999 or searchSiteID == 863 or (867 <= searchSiteID <= 870):
+                    results = PAsearchSites.networkPuffy.search(results, encodedTitle, title, searchTitle, siteNum, lang, searchByDateActor, searchDate, searchSiteID)
+
+            ###############
+            ## SinX
+            ###############
+            if siteNum == 864:
+                if searchSiteID == 9999 or (864 <= searchSiteID <= 866) or searchSiteID == 871:
+                    results = PAsearchSites.networkSinX.search(results, encodedTitle, title, searchTitle, siteNum, lang, searchByDateActor, searchDate, searchSiteID)
+                    
+            ###############
+            ## Kinky Spa
+            ###############
+            if siteNum == 872:
+                if searchSiteID == 9999 or searchSiteID == 872:
+                    results = PAsearchSites.network1service.search(results, encodedTitle, title, searchTitle, siteNum, lang, searchByDateActor, searchDate, searchSiteID)
+
             siteNum += 1
 
         results.Sort('score', descending=True)
@@ -1417,7 +1466,7 @@ class PhoenixAdultAgent(Agent.Movies):
         ##   Gamma Entertainment                                    ##
         ##                                                          ##
         ##############################################################
-        if siteID == 183 or siteID == 278 or siteID == 281 or (siteID >= 285 and siteID <= 287) or (siteID >= 329 and siteID <= 330) or (siteID >= 351 and siteID <= 360) or (siteID >= 365 and siteID <= 374) or (siteID >= 381 and siteID <= 392) or (siteID >= 460 and siteID <= 466) or siteID == 750 or siteID == 843:
+        if siteID == 183 or siteID == 278 or siteID == 281 or (siteID >= 285 and siteID <= 287) or (siteID >= 329 and siteID <= 330) or (siteID >= 351 and siteID <= 360) or (siteID >= 365 and siteID <= 374) or (siteID >= 381 and siteID <= 392) or (siteID >= 460 and siteID <= 466) or siteID == 750 or siteID == 843 or siteID == 861:
             metadata = PAsearchSites.networkGammaEnt.update(metadata, siteID, movieGenres, movieActors)
 
         ##############################################################
@@ -2194,6 +2243,54 @@ class PhoenixAdultAgent(Agent.Movies):
         ##                                                          ##
         ##############################################################
         if siteID == 841:
+            metadata = PAsearchSites.network1service.update(metadata, siteID, movieGenres, movieActors)
+
+        ##############################################################
+        ##                                                          ##
+        ##   Deviant Hardcore                                       ##
+        ##                                                          ##
+        ##############################################################
+        if siteID == 859:
+            metadata = PAsearchSites.network1service.update(metadata, siteID, movieGenres, movieActors)
+
+        ##############################################################
+        ##                                                          ##
+        ##   She Will Cheat                                         ##
+        ##                                                          ##
+        ##############################################################
+        if siteID == 860:
+            metadata = PAsearchSites.network1service.update(metadata, siteID, movieGenres, movieActors)
+
+        ##############################################################
+        ##                                                          ##
+        ##   SinsLife                                               ##
+        ##                                                          ##
+        ##############################################################
+        if siteID == 862:
+            metadata = PAsearchSites.siteSinsLife.update(metadata, siteID, movieGenres, movieActors)
+
+        ##############################################################
+        ##                                                          ##
+        ##   Puffy Network                                          ##
+        ##                                                          ##
+        ##############################################################
+        if siteID == 863 or (siteID >=  867 and siteID <= 870):
+            metadata = PAsearchSites.networkPuffy.update(metadata, siteID, movieGenres, movieActors)
+
+        ##############################################################
+        ##                                                          ##
+        ##   SinX                                                   ##
+        ##                                                          ##
+        ##############################################################
+        if siteID == 871 or (siteID >=  864 and siteID <= 866):
+            metadata = PAsearchSites.networkSinX.update(metadata, siteID, movieGenres, movieActors)
+
+        ##############################################################
+        ##                                                          ##
+        ##   Kinky Spa                                              ##
+        ##                                                          ##
+        ##############################################################
+        if siteID == 872:
             metadata = PAsearchSites.network1service.update(metadata, siteID, movieGenres, movieActors)
 
         ##############################################################