+++ Contents/Code/siteSinsLife.py
@@ -6,20 +6,19 @@ def search(results,encodedTitle,title,searchTitle,siteNum,lang,searchByDateActor
     if searchSiteID != 9999:
         siteNum = searchSiteID
     searchResults = HTML.ElementFromURL(PAsearchSites.getSearchSearchURL(siteNum) + encodedTitle)
-    for searchResult in searchResults.xpath('//div[@style="position:relative; background:black;"]'):
-        titleNoFormatting = searchResult.xpath('.//a')[0].get('title')
+    for searchResult in searchResults.xpath('//div[4]/div/div[3]/div/div'):
+        titleNoFormatting = searchResult.xpath('.//a')[0].get('title').title()
         curID = searchResult.xpath('.//a')[0].get('href').replace('/','_').replace('?','!')
-        releaseDate = parse(searchResult.xpath('.//span[@class="date"]')[0].text_content().strip()).strftime('%Y-%m-%d')
         score = 100 - Util.LevenshteinDistance(searchTitle.lower(), titleNoFormatting.lower())
-        results.Append(MetadataSearchResult(id = curID + "|" + str(siteNum), name = titleNoFormatting + " [VIPissy] " + releaseDate, score = score, lang = lang))
+        results.Append(MetadataSearchResult(id = curID + "|" + str(siteNum), name = titleNoFormatting + " [SinsLife] ", score = score, lang = lang))
 
     return results
 
 def update(metadata,siteID,movieGenres,movieActors):
     Log('******UPDATE CALLED*******')
 
-    temp = str(metadata.id).split("|")[0].replace('_','/').replace('?','!')
-    url = 'https://www.vipissy.com' + temp
+    temp = str(metadata.id).split("=")[-1].replace('_','/').replace('?','!')
+    url = 'https://nats.thesinslife.com/signup/signup.php?&setID=' + temp
     detailsPageElements = HTML.ElementFromURL(url)
     art = []
     metadata.collections.clear()
@@ -27,16 +26,16 @@ def update(metadata,siteID,movieGenres,movieActors):
     movieActors.clearActors()
 
     # Studio
-    metadata.studio = 'VIPissy'
+    metadata.studio = 'SinsLife'
 
     # Title
-    metadata.title = detailsPageElements.xpath('//section[@class="downloads"]/strong')[0].text_content().strip()
+    metadata.title = detailsPageElements.xpath('//div[4]/div/div[2]/div/div/div[2]/div[1]/div/h3')[0].text_content().strip().title()
 
     # Summary
-    all_summary = detailsPageElements.xpath('//div/section[4]/div')[0].text_content().strip()
-    tags_summary = detailsPageElements.xpath('//div/section[4]/div/p')[0].text_content().strip()
-    summary = all_summary.replace(tags_summary, '')
-    summary = summary.split("Show more...")[0].strip()
+    summary = detailsPageElements.xpath('//div[4]/div/div[2]/div/div/div[2]/div[2]/p')[0].text_content().strip()
+#    tags_summary = detailsPageElements.xpath('//div/section[4]/div/p')[0].text_content().strip()
+#    summary = all_summary.replace(tags_summary, '')
+#    summary = summary.split("Show more...")[0].strip()
     metadata.summary = summary
     # Log("Summary:" + metadata.summary)
     
@@ -46,22 +45,22 @@ def update(metadata,siteID,movieGenres,movieActors):
     metadata.collections.add(tagline)
 
     # Genres
-    genres = detailsPageElements.xpath('//div/section[4]/div/p/a')
-    if len(genres) > 0:
-        for genreLink in genres:
-            genreName = genreLink.text_content().strip().lower()
-            movieGenres.addGenre(genreName)
+#    genres = detailsPageElements.xpath('//div/section[4]/div/p/a')
+#    if len(genres) > 0:
+#        for genreLink in genres:
+#            genreName = genreLink.text_content().strip().lower()
+#            movieGenres.addGenre(genreName)
 
     # Release Date
-    date = detailsPageElements.xpath('//div/section[2]/dl/dd[2]')[0].text_content().strip()
+    date = detailsPageElements.xpath('//div[4]/div/div[2]/div/div/div[2]/div[1]/div/div[1]')[0].text_content().strip("Release Date:")
     Log("Date:" + date)
     if len(date) > 0:
-        date_object = datetime.strptime(date, '%b %d, %Y')
+        date_object = datetime.strptime(date, '%B %d, %Y')
         metadata.originally_available_at = date_object
         metadata.year = metadata.originally_available_at.year
 
     # Actors
-    actors = detailsPageElements.xpath('//div/section[2]/dl/dd[1]/a')
+    actors = detailsPageElements.xpath('//div[4]/div/div[2]/div/div/div[2]/div[3]/ul/li/a/..')
     if len(actors) > 0:
         if len(actors) == 3:
             movieGenres.addGenre("Threesome")
@@ -71,16 +70,14 @@ def update(metadata,siteID,movieGenres,movieActors):
             movieGenres.addGenre("Orgy")
         for actorLink in actors:
             actorName = str(actorLink.text_content().strip())
-            try:
-                actorPageURL = actorLink.get("href")
-                if 'http' not in actorPageURL:
-                    actorPageURL = PAsearchSites.getSearchBaseURL(siteID) + actorPageURL
-                actorPage = HTML.ElementFromURL(actorPageURL)
-                actorPhotoURL = actorPage.xpath('//div/section[1]/div/div[1]/img')[0].get("src")
-                if 'http' not in actorPhotoURL:
-            	    actorPhotoURL = PAsearchSites.getSearchBaseURL(siteID) + actorPhotoURL
-            except:
-                actorPhotoURL = ""
+            #actorPageURL = actorLink.get("src")
+    # can't get this to work, below is temp fix
+            if len(actors) == 1:
+                actorPageURL = detailsPageElements.xpath('//div[4]/div/div[2]/div/div/div[2]/div[3]/ul/li/a/img[@class="category_model_thumb stdimage thumbs target"]')[0].get("src")
+                actorPageURL = 'https:' + actorPageURL
+                actorPhotoURL = actorPageURL
+            if len(actors) > 1:
+                actorPhotoURL = ''
             movieActors.addActor(actorName,actorPhotoURL)
 
     # Director
@@ -89,22 +86,11 @@ def update(metadata,siteID,movieGenres,movieActors):
 
     # Video trailer background image
     try:
-        temp2 = str(metadata.id).split("|")[0].replace('_','/').replace('?','!')
-        temp2 = temp2.split("/updates")[1]
-        twitterBG = 'https://media.vipissy.com/videos' + temp2 + 'cover/l.jpg'
+        twitterBG = detailsPageElements.xpath('//div[4]/div/div[2]/div/div/div[1]/div[2]/div/div[1]/img')[0].get('src')
+        twitterBG = 'https:' + twitterBG
         art.append(twitterBG)
     except:
         pass
-
-    # Posters
-    posters = detailsPageElements.xpath('//div[contains(@id, "pics2")]//div//ul//li//div//div//img')
-    posterNum = 1
-    Log(str(len(posters)))
-    for poster in posters:
-        posterURL = poster.get("src")
-        metadata.posters[posterURL] = Proxy.Preview(HTTP.Request(posterURL, headers={'Referer': 'http://www.google.com'}).content, sort_order = posterNum)
-        posterNum += 1
-    
 
     j = 1
     Log("Artwork found: " + str(len(art)))
@@ -123,8 +109,6 @@ def update(metadata,siteID,movieGenres,movieActors):
                 if width > 100 and width > height:
                     # Item is an art item
                     metadata.art[posterUrl] = Proxy.Preview(HTTP.Request(posterUrl, headers={'Referer': 'http://www.google.com'}).content, sort_order = j)
-                    # and Item is a poster
-                    metadata.posters[posterUrl] = Proxy.Preview(HTTP.Request(posterUrl, headers={'Referer': 'http://www.google.com'}).content, sort_order = j)
                 j = j + 1
             except:
                 pass